+++ src/structs/Probe.jl
@@ -4,5 +4,5 @@ abstract type AbstractProbe end
 
 @kwdef mutable struct Probe <: AbstractProbe
     origin::Point = Point()
-    geometry::Array{Float64, 2} = Array{Float64, 2}(undef, 0, 7)
+    geometry::Array{Float32, 2} = Array{Float32, 2}(undef, 0, 7)
 end
